Feat/show quotes and add task modal button
메인페이지 내에 헤더에 명언 api 호출과 태스트 작성 버튼 생성
