+++ src/components/Main/DailyCalender.tsx
@@ -1,13 +1,12 @@
-import React, { useEffect } from "react";
 import { styled } from "styled-components";
 import { Tasks, getFollowersTasks, getMyTasks } from "../../api/tasks";
 import { useQuery } from "@tanstack/react-query";
 import { Followers, getFollowers } from "../../api/users";
+import { useModalStore } from "../../config/useModalStore";
+import AddTaskModal from "../modal/AddTaskModal";
+import { getQuotes } from "../../api/getQuotes";
 
 const DailyCalender = () => {
-  // 1. 현재 로그인된 유저의 uid를 가지고 와서 uid와 현재 날짜를 이용해 task들을 가져오기
-  // 2. 현재 로그인된 유저가 팔로우하는 사람들의 목록 가져오기
-  // 3. 팔로우한 사람들의 uid와 현재 날짜를 이용해 팔로워들의 task 가져오기
   const today = new Date().toISOString().slice(0, 10);
   const myId = "ae06168e-38d9-4a05-a2d6-41e5c0a7aac6";
 
@@ -27,7 +26,6 @@ const DailyCalender = () => {
         followers?.map((follower: Followers) => follower.from),
     }
   );
-  console.log("팔로워들", followers);
 
   const { data: followersTasks } = useQuery(
     ["followersTasks"],
@@ -38,72 +36,106 @@ const DailyCalender = () => {
     { enabled: !!followers }
   );
 
+  const { data: quotes } = useQuery(["quotes"], async () => {
+    const quotesData = await getQuotes();
+    return quotesData;
+  });
+
+  console.log(quotes);
+
+  const { addTaskModalVisible, changeAddTaskModalstatus } = useModalStore();
+
   return (
-    <S.CalenderContainer>
-      <S.TimeStampContainer>
-        {Array.from({ length: 25 }, (_, i) => (
-          <S.TimeStamp key={i}>{`${i}:00`}</S.TimeStamp>
-        ))}
-      </S.TimeStampContainer>
-      <S.TaskContainer>
-        {tasks &&
-          tasks.map((task) => {
-            const endHour = task.end_time;
-            const startHour = task.start_time;
-            const height = (endHour - startHour) * 80;
-            const top = startHour * 80;
-            return (
-              <S.TaskBox height={height} top={top}>
-                <S.Task>
-                  <p>{task.title}</p>
-                </S.Task>
-              </S.TaskBox>
-            );
-          })}
-      </S.TaskContainer>
-      {followers &&
-        followers.map((follower: string) => {
-          console.log("팔로워", follower);
-          const followerTasks = followersTasks?.filter(
-            (followersTask) => followersTask.user_id === follower
-          );
-          console.log("팔로워의 태스크들", followerTasks);
-          return (
-            <S.TaskContainer>
-              {followerTasks?.map((followerTask: Tasks) => {
-                console.log("팔로워의 태스크", followerTask);
-                const endHour = followerTask.end_time;
-                const startHour = followerTask.start_time;
+    <>
+      {addTaskModalVisible ? <AddTaskModal todayDefault={true} /> : null}
+      <S.Header>
+        <div>{quotes?.advice}</div>
+        <button onClick={changeAddTaskModalstatus}>버튼</button>
+      </S.Header>
+      <S.Container>
+        <S.CalenderContainer>
+          <S.TimeStampContainer>
+            <S.TimeStamp>Time</S.TimeStamp>
+            {Array.from({ length: 25 }, (_, i) => (
+              <S.TimeStamp key={i}>{`${i}:00`}</S.TimeStamp>
+            ))}
+          </S.TimeStampContainer>
+          <S.TaskContainer>
+            <S.TaskBox>My Task</S.TaskBox>
+            {tasks &&
+              tasks.map((task) => {
+                const endHour = task.end_time;
+                const startHour = task.start_time;
                 const height = (endHour - startHour) * 80;
-                const top = startHour * 80;
+                const top = (startHour + 1) * 80;
                 return (
                   <S.TaskBox height={height} top={top}>
                     <S.Task>
-                      <p>{followerTask.title}</p>
+                      <p>{task.title}</p>
                     </S.Task>
                   </S.TaskBox>
                 );
               })}
-            </S.TaskContainer>
-          );
-        })}
-    </S.CalenderContainer>
+          </S.TaskContainer>
+        </S.CalenderContainer>
+        <S.FollowersCalenderContainer>
+          {followers &&
+            followers.map((follower: string) => {
+              const followerTasks = followersTasks?.filter(
+                (followersTask) => followersTask.user_id === follower
+              );
+              return (
+                <S.TaskContainer>
+                  {followerTasks?.map((followerTask: Tasks) => {
+                    const endHour = followerTask.end_time;
+                    const startHour = followerTask.start_time;
+                    const height = (endHour - startHour) * 80;
+                    const top = (startHour + 1) * 80;
+                    return (
+                      <>
+                        <S.TaskBox>{followerTask.user_id}</S.TaskBox>
+                        <S.TaskBox height={height} top={top}>
+                          <S.Task>
+                            <p>{followerTask.title}</p>
+                          </S.Task>
+                        </S.TaskBox>
+                      </>
+                    );
+                  })}
+                </S.TaskContainer>
+              );
+            })}
+        </S.FollowersCalenderContainer>
+      </S.Container>
+    </>
   );
 };
 
 export default DailyCalender;
 
 interface styleProps {
-  height: number;
-  top: number;
+  height?: number;
+  top?: number;
 }
 
 const S = {
+  Header: styled.div`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+  `,
+  Container: styled.div`
+    display: flex;
+    flex-direction: row;
+  `,
   CalenderContainer: styled.div`
     display: flex;
     flex-direction: row;
     background-color: azure;
+    padding: 10px;
   `,
+
   TimeStampContainer: styled.div`
     background-color: antiquewhite;
     margin: 0 30px;
@@ -113,9 +145,16 @@ const S = {
   `,
   TaskContainer: styled.div`
     background-color: beige;
-    width: 400px;
+    min-width: 400px;
     position: relative;
-    margin: 10px;
+    margin: 0 10px;
+  `,
+  FollowersCalenderContainer: styled.div`
+    display: flex;
+    flex-direction: row;
+    background-color: azure;
+    overflow-x: scroll;
+    padding: 10px;
   `,
   TaskBox: styled.div<styleProps>`
     height: ${(props) => props.height}px;
