+++ src/components/Main/ExplorePeople.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ExplorePeople = () => {
+  return <div>ExplorePeople</div>;
+};
+
+export default ExplorePeople;
