+++ src/config/useModalStore.tsx
@@ -1,0 +1,12 @@
+import { create } from "zustand";
+
+type ModalStore = {
+  addTaskModalVisible: boolean;
+  changeAddTaskModalstatus: () => void;
+};
+
+export const useModalStore = create<ModalStore>()((set) => ({
+  addTaskModalVisible: false,
+  changeAddTaskModalstatus: () =>
+    set((state) => ({ addTaskModalVisible: !state.addTaskModalVisible })),
+}));
