+++ src/pages/Main.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import DailyCalender from "../components/Main/DailyCalender";
+import DailyCalender from "../components/main/DailyCalender";
 
 const Main = () => {
   return (
